requirements.txt
pyfastx==0.8.4
PyYAML==6.0
